Fix parlay record stats query crashing on PostgreSQL
PostgreSQL's ROUND(value, precision) only accepts `numeric` type, not
`double precision`. The AVG() expressions return double precision which
caused a 500 error: "function round(double precision, integer) does not
exist". Added ::numeric casts to match the pattern already used in
copilot.py.

+++ backend/src/agents/parlay_tracker.py
@@ -359,8 +359,8 @@ async def get_parlay_record(db: AsyncSession, days: int = 30) -> dict:
                 COUNT(*) FILTER (WHERE result IS NOT NULL)          AS total,
                 COUNT(*) FILTER (WHERE result = 'win')              AS wins,
                 COUNT(*) FILTER (WHERE result = 'loss')             AS losses,
-                ROUND(AVG(legs_hit::float / NULLIF(legs_total,0)) * 100, 1) AS avg_legs_hit_pct,
-                ROUND(AVG(combined_prob) * 100, 1)                  AS avg_model_prob_pct
+                ROUND((AVG(legs_hit::float / NULLIF(legs_total,0)) * 100)::numeric, 1) AS avg_legs_hit_pct,
+                ROUND((AVG(combined_prob) * 100)::numeric, 1)                  AS avg_model_prob_pct
             FROM daily_parlays
             WHERE game_date >= CURRENT_DATE - (CAST(:days AS INTEGER) * INTERVAL '1 day')
             GROUP BY parlay_name
